requirements.txt
# Core RL framework
gymnasium>=0.29.0
stable-baselines3>=2.1.0
sb3-contrib>=2.1.0

# Deep Learning & GNN
torch>=2.0.0
torch-geometric>=2.4.0
torch-scatter>=2.1.0
torch-sparse>=0.6.0

# Data & Config
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0
omegaconf>=2.3.0

# Monitoring & Logging
tensorboard>=2.14.0
wandb>=0.15.0
tqdm>=4.65.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
networkx>=3.1

# MQTT Integration (Option 4)
paho-mqtt>=1.6.1

# Testing & Quality
pytest>=7.4.0
black>=23.7.0
flake8>=6.1.0

# Java Bridge (optional, if calling iFogSim from Python)
py4j>=0.10.9.7
